requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
djangorestframework==3.16.1
drf-yasg==1.21.11
greenlet==3.2.4
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
packaging==25.0
playwright==1.55.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
